+++ OOP_Project_1/OOP_Project_1.vcxproj
@@ -139,12 +139,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="main_menu.cpp" />
     <ClCompile Include="Management.cpp" />
     <ClCompile Include="Sorting.cpp" />
     <ClCompile Include="student.cpp" />
     <ClCompile Include="student_search.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="main_menu.h" />
     <ClInclude Include="Management.h" />
     <ClInclude Include="insertion.cpp" />
     <ClInclude Include="insertion.h" />
