+++ OOP_Project_1/Sorting.cpp
@@ -1,18 +1,18 @@
-#include <iostream>
-#include "sort.h"
-#include "student_search.h"
-#include "student.h"
-#include "Management.h"
-
-using namespace std;
-
-
-
-
-int main() {
-	//Management management;
-	//management.sort_menu();
-	Management m;
-	m.initial_get_info();
-	m.Search();
-}
+//#include <iostream>
+//#include "sort.h"
+//#include "student_search.h"
+//#include "student.h"
+//#include "Management.h"
+//
+//using namespace std;
+//
+//
+//
+//
+//int main() {
+//	//Management management;
+//	//management.sort_menu();
+//	Management m;
+//	m.initial_get_info();
+//	m.Search();
+//}
