+++ OOP_Project_1/main_menu.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+#include "Management.h"5
+
+using namespace std;
+
+int main() {
+	Management management;
+	management.menu();
+}
