+++ OOP_Project_1/student.cpp
@@ -1,42 +1,42 @@
-#include<iostream>
-#include "student.h"
-
-using namespace std;
-
-	
-Student::Student() {};
-Student::Student(string name, string id, string dept, string birth, string tel) {
-	Name = name;
-	ID = id;
-	Dept = dept;
-	Birth = birth;
-	Tel = tel;
-}
-
-void Student:: printInfo() {
-	// cout << Name << "\n" << ID << "\n" << Dept << "\n" << Birth << "\n" << Tel << "\n";
-	cout << Name << " " << ID << " " << Birth << " " << Dept << " " << Tel << " " << endl;
-}
-
-string Student:: getName()  {
-	return Name;
-}
-
-string Student::getID()  {
-	return ID;
-}
-
-string Student:: getDept()  {
-	return Dept;
-}
-
-string Student:: getBirth()  {
-	return Birth;
-}
-
-string Student:: getTel()  {
-	return Tel;
-}
-
-
-
+//#include<iostream>
+//#include "student.h"
+//
+//using namespace std;
+//
+//	
+//Student::Student() {};
+//Student::Student(string name, string id, string dept, string birth, string tel) {
+//	Name = name;
+//	ID = id;
+//	Dept = dept;
+//	Birth = birth;
+//	Tel = tel;
+//}
+//
+//void Student:: printInfo() {
+//	// cout << Name << "\n" << ID << "\n" << Dept << "\n" << Birth << "\n" << Tel << "\n";
+//	cout << Name << " " << ID << " " << Birth << " " << Dept << " " << Tel << " " << endl;
+//}
+//
+//string Student:: getName()  {
+//	return Name;
+//}
+//
+//string Student::getID()  {
+//	return ID;
+//}
+//
+//string Student:: getDept()  {
+//	return Dept;
+//}
+//
+//string Student:: getBirth()  {
+//	return Birth;
+//}
+//
+//string Student:: getTel()  {
+//	return Tel;
+//}
+//
+//
+//
