+++ OOP_Project_1/student.h
@@ -12,18 +12,37 @@ class Student {
 	string Tel;
 
 public:
-	Student();
-	Student(string name, string id, string dept, string birth, string tel);
+	Student(){};
+	Student(string name, string id, string dept, string birth, string tel) {
+		this->Name = name;
+		this->ID = id;
+		this->Dept = dept;
+		this->Birth = birth;
+		this->Tel = tel;
+	};
 
-	void printInfo();
+	void printInfo() {
+		// cout << Name << "\n" << ID << "\n" << Dept << "\n" << Birth << "\n" << Tel << "\n";
+		cout << Name << " " << ID << " " << Birth << " " << Dept << " " << Tel << " " << endl;
+	}
 
-	string getName();
+	string getName() {
+		return Name;
+	}
 
-	string getID() ;
+	string getID() {
+		return ID;
+	}
 
-	string getDept() ;
+	string getDept() {
+		return Dept;
+	}
 
-	string getBirth() ;
+	string getBirth() {
+		return Birth;
+	}
 
-	string getTel() ;
+	string getTel() {
+		return Tel;
+	}
 };
